Create Exe3
When 'NumProducer' and 'NumConsumer' are set to distinct values, the corresponding number of producer and consumer threads will be created. If 'NumConsumer' is set to 3 and 'NumProducer' is set to 2, for instance, the code will generate 2 producer threads and 3 consumer threads.

When one of them is set to "0", no threads of that type are created. For instance, if 'NumProducer' is set to 0 and 'NumConsumer' is set to 3, only 3 consumer threads will be created.